arm64: Improve CNTVCT/CNTFRQ macros for compat read
Having opcode checked might not be sufficient in some corner cases
as an MRC or MRRC instruction might jump into the same trap handler
while having the same opcode but reading a different corprocessor.
In this case, that instruction will get an unexpected result -- a
CNTVCT or CNTFRQ value.

So this patch just simply check all fields (except condition fields)
to make sure the instruction is for CNTVCT or CNTFRQ. Since it does
not hurt to unconditionally read a CNTVCT/CNTFRQ value, this change
just ignore the condition field and read it anyway.

Bug 2044346

Change-Id: Ia63855573d1b63d5bf503567556bd64bd83e0458
Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1637504
Reviewed-by: Vinayak Pane <vpane@nvidia.com>
Tested-by: Vinayak Pane <vpane@nvidia.com>